requirements.txt
Django==3.2.14
Pillow==9.4.0
Six==1.16.0
django-bootstrap-modal-forms==2.2.0
